Mosaic_python/requirements.txt
Pillow
ttkbootstrap
scipy
numpy
groq
python-dotenv
